Improve protocol used to receive frames from WM

+++ async-toolkit/m3utils/mby-merged/wm/src/client/c/mbay_dpi_client.h
@@ -50,12 +50,12 @@ int wm_disconnect(void);
 int wm_reg_write(const uint32_t addr, const uint64_t val);
 int wm_reg_read(const uint32_t addr, uint64_t *val);
 
-int wm_pkt_push(int port, const uint8_t *data, uint32_t len);
-int wm_pkt_get(int *port, uint8_t *data, uint32_t *len);
+int wm_pkt_push(uint16_t port, const uint8_t *data, uint32_t len);
+int wm_pkt_get(uint16_t *port, uint8_t *data, uint32_t *len);
 
 #ifndef NO_SV
-int wm_svpkt_push(int port, const svOpenArrayHandle sv_data, uint32_t len);
-int wm_svpkt_get(int *port, svOpenArrayHandle sv_data, uint32_t *len);
+int wm_svpkt_push(uint16_t  port, const svOpenArrayHandle sv_data, uint32_t len);
+int wm_svpkt_get(uint16_t  *port, svOpenArrayHandle sv_data, uint32_t *len);
 #endif
 
 #endif /* __MBAY_DPI_CLIENT_H_ */
